Use WPARAM as uint64_t and LPARAM as int64_t

+++ atom/browser/native_window.cc
@@ -469,7 +469,8 @@ void NativeWindow::NotifyWindowMessage(UINT message, WPARAM w_param,
                                        LPARAM l_param) {
   FOR_EACH_OBSERVER(
       NativeWindowObserver, observers_,
-      OnWindowMessage(message, w_param, static_cast<uint64_t>(l_param)));
+      OnWindowMessage(message, w_param,
+                      l_param));
 }
 #endif
